Correct switch statements
Remove unnecessary default case in switch statement.  When strict =
true in tsconfig, there is no need for the default clause in the switch
statement to achieve compile-time exhaustive checking.  This is true as
long as the function defines a non-void return type.
The checkNil function still needs the default case, because the return
type in void.