Refine return type of next() for better type-checking
Change next(), stop(), loop() to return 'done' or Promise<'done'>
instead of returning void.  This forces the compiler to do flow typing
and trigger exhaustiveness checking for the Result ADT kind
discrimination in the next() switch.  Otherwise, if next() returns void,
the compiler doesn't care that the switch is not exhaustive, because
falling through the switch would return void anyway.